fix: 批次建帳交易隔離 + Stats/AR bigint 型態 + 聊天記錄排序穩定性
- auth/batch-create: SAVEPOINT 隔離單筆失敗（原本一筆違反 unique 會讓整批 COMMIT 靜默變 ROLLBACK，API 回「created」但實際 DB 沒進）
- stats/overview: COUNT/SUM 加 ::int，避免 pg bigint 轉字串讓前端 Dashboard 算術失誤
- ar-practice/students: 同上型態修正
- chat/stream + /history: ORDER BY 補 id tie-breaker，避免同毫秒訊息順序亂跳造成 Sliding Window 錯序
- users/usage: questionLimit 過期常數 50 → 30 對齊新 rate limit

+++ server/routes/ar-practice.js
@@ -94,8 +94,8 @@ router.get('/students', async (req, res) => {
   try {
     const { rows } = await pool.query(`
       SELECT u.id, u.student_id, u.name,
-        COUNT(s.id) as total_sessions,
-        COUNT(CASE WHEN s.status = 'COMPLETED' THEN 1 END) as completed_sessions,
+        COUNT(s.id)::int as total_sessions,
+        COUNT(CASE WHEN s.status = 'COMPLETED' THEN 1 END)::int as completed_sessions,
         MAX(s.created_at) as last_practice,
         COALESCE(AVG(s.duration_seconds), 0)::int as avg_duration
       FROM users u
